+++ src/candlewick/multibody/RobotScene.cpp
@@ -226,8 +226,8 @@ void RobotScene::initCompositePipeline(const MeshLayout &layout) {
 
   pipelines.wboitComposite = SDL_CreateGPUGraphicsPipeline(device(), &desc);
   if (!pipelines.wboitComposite) {
-    terminate_with_message(
-        std::format("Failed to create WBOIT pipeline: %s", SDL_GetError()));
+    terminate_with_message("Failed to create WBOIT pipeline: %s",
+                           SDL_GetError());
   }
 }